+++ TabbedPanelMod/src/main/groovy/TPM.groovy
@@ -1,6 +1,6 @@
 package edofro.tabbedpanelmod
 
-import javax.swing.JButton
+
 import java.awt.Dimension
 import java.awt.Insets
 
@@ -15,6 +15,7 @@ import org.freeplane.plugin.script.proxy.ScriptUtils
 
 
 class TPM{
+    // region properties
     static final c = ScriptUtils.c()
     static final JTabbedPane FPTabPane = ui.getFreeplaneTabbedPanel()
 
@@ -35,12 +36,29 @@ class TPM{
 
     static boolean savingTpmProps = false
 
+    // endregion
+
+    // region methods
+
     def static modifyTab(int i,  iconForTab = null){
         if(isModActive()) {
-            def toolTip = FPTabPane.getTitleAt(i) ?: FPTabPane.getToolTipTextAt(i)
-            FPTabPane.setToolTipTextAt(i, toolTip)
+            def tabName = FPTabPane.getTitleAt(i) ?: FPTabPane.getToolTipTextAt(i)
+            println('modifyTab: tabName ' + tabName)
+            println('modifyTab: iconForTab ' + iconForTab)
+            println('modifyTab: FPTabPane.tabIconNames ' + FPTabPane.tabIconNames)
+            if(iconForTab){
+                FPTabPane.tabIconNames[tabName] = iconForTab
+            } else {
+                iconForTab = FPTabPane.tabIconNames[tabName]
+            }
+            println('modifyTab: iconForTab ' + iconForTab)
+            println('modifyTab: FPTabPane.tabIconNames ' + FPTabPane.tabIconNames)
+            FPTabPane.setToolTipTextAt(i, tabName)
             FPTabPane.setTitleAt(i, null)
-            FPTabPane.setTabComponentAt(i, tabButton(i, toolTip, iconForTab))
+            FPTabPane.setTabComponentAt(i, tabButton(i, tabName, iconForTab))
+            def w = FPTabPane.tabWidths[tabName]
+            if(w)
+                resizeTP(w)
         }
     }
 
@@ -48,7 +66,7 @@ class TPM{
         FPTabPane.hasProperty('isModded')? FPTabPane.isModded : false
     }
 
-    def static tabButton(int indexP, String toolTip,  iconForTab = null){
+    def static tabButton(int indexP, String tabName,  iconForTab = null){
         def TPMaction = {e ->
             collapsedWidth = e.source.parent.width
             def botones = e.source.parent.components
@@ -62,26 +80,29 @@ class TPM{
                 resizeTP(collapsedWidth)
             }else{
                 FPTabPane.setSelectedIndex(index)
-                if(FPTabPane.tabWidths[toolTip]){
-                    resizeTP(FPTabPane.tabWidths[toolTip])
+                if(FPTabPane.tabWidths[tabName]){
+                    resizeTP(FPTabPane.tabWidths[tabName])
                 } else if (FPTabPane.collapsed){
                     resizeTP(FPTabPane.originalWidth)
                 } else {
-                    FPTabPane.tabWidths[toolTip] = FPTabPane.width
+                    FPTabPane.tabWidths[tabName] = FPTabPane.width
                 }
                 FPTabPane.collapsed = false
             }
         }
         def icono = iconForTab?menuUtils.getMenuItemIcon('IconAction.' + iconForTab):null
-        icono ?= getIcon(indexP, toolTip)
+        icono ?= getIconFromTPM(indexP, tabName)
         def btn = swingBuilder.button(
                 horizontalAlignment : SwingConstants.LEFT,
                 icon                : icono,
-                toolTipText         : toolTip,
-                margin              : new Insets(0,0,0,0),
-                borderPainted       : false,
-                opaque              : false,
-                actionPerformed     : TPMaction
+                toolTipText         : tabName,
+                margin              : new Insets(5,5,5,5),
+              //  borderPainted       : true,
+              //  opaque              : true,
+              //  background          : Color.CYAN,
+                actionPerformed     : TPMaction,
+             //   minimumSize         : new Dimension(40,10),
+             //   preferredSize       : new Dimension(40,40),
         )
         return btn
     }
@@ -95,7 +116,7 @@ class TPM{
         parentBox.repaint()
     }
 
-    def static getIcon(i,t){
+    def static getIconFromTPM(i, t){
         def ico = 'IconAction.TabbedPanelMod/'
         menuUtils.getMenuItemIcon("${ico}${t}")?:menuUtils.getMenuItemIcon("${ico}${t.toLowerCase()}")?: menuUtils.getMenuItemIcon("${ico}${String.format('%02d',i)}")?: menuUtils.getMenuItemIcon(iconos[i])
     }
@@ -107,7 +128,7 @@ class TPM{
             tpmProps.each{k,v ->
                 tw[k] = ((String) v).isInteger()?v.toInteger():v
             }
-            println tw
+            // println tw
         }
         return tw
     }
@@ -115,7 +136,7 @@ class TPM{
     def static saveTabWidths(){
         if (!savingTpmProps && FPTabPane.hasProperty('tabWidths')) {
             savingTpmProps = true
-            new Timer().runAfter(2000) {
+            new Timer().runAfter(10000) {
                 def map2 = FPTabPane.tabWidths.clone()
                 map2.each { k, v -> map2[k] = v.toString() }
                 if (tpmProps != map2) {
@@ -131,4 +152,6 @@ class TPM{
         }
 
     }
+
+    // endregion
 }