+++ TabbedPanelMod/history.md
@@ -1,5 +1,11 @@
 # History
 
+## v0.5
+
+* Now when adding a tab that was used a time before, it automatically changes its width to the last one saved
+* bigger Tab buttons and narrower tabs container
+* Now it remembers the icons when going to TabsOnTop and back
+
 ## v0.4
 
 * added boolean property "isModded" to TabPanel